Refactoring code to multiple files and package into 'pyhindsight' for PyPI
